Move babel-standalone's build scripts to the repo root.
