--- packages/babel-standalone/Gulpfile.js
@@ -1,6 +1,0 @@
-const registerBabelStandaloneTask = require("./src/gulpTasks")
-  .registerBabelStandaloneTask;
-const gulp = require("gulp");
-
-const version = require("./package.json").version;
-registerBabelStandaloneTask(gulp, "babel", "Babel", __dirname, version);
