+++ packages/babel-standalone/package.json
@@ -10,8 +10,6 @@
   ],
   "devDependencies": {
     "babel-core": "7.0.0-beta.2",
-    "babel-helper-builder-react-jsx": "7.0.0-beta.2",
-    "babel-loader": "7.1.1",
     "babel-plugin-check-es2015-constants": "7.0.0-beta.2",
     "babel-plugin-external-helpers": "7.0.0-beta.2",
     "babel-plugin-inline-replace-variables": "1.3.1",
@@ -97,16 +95,7 @@
     "babel-preset-stage-1": "7.0.0-beta.2",
     "babel-preset-stage-2": "7.0.0-beta.2",
     "babel-preset-stage-3": "7.0.0-beta.2",
-    "babel-preset-typescript": "7.0.0-beta.2",
-    "gulp-rename": "^1.2.2",
-    "gulp-uglify": "^3.0.0",
-    "json-loader": "^0.5.7",
-    "lazypipe": "^1.0.1",
-    "pump": "^1.0.2",
-    "react": "^16.0.0-beta.3",
-    "webpack": "^3.4.1",
-    "webpack-dependency-suite": "^2.4.4",
-    "webpack-stream": "^4.0.0"
+    "babel-preset-typescript": "7.0.0-beta.2"
   },
   "keywords": [
     "babel",
@@ -120,11 +109,5 @@
   "bugs": {
     "url": "https://github.com/babel/babel-standalone/issues"
   },
-  "homepage": "https://github.com/babel/babel-standalone#readme",
-  "jest": {
-    "transformIgnorePatterns": [
-      "/node_modules/",
-      "babel.js"
-    ]
-  }
+  "homepage": "https://github.com/babel/babel-standalone#readme"
 }
